tests: fix up fixtures and dependencies
