+++ test-reqs.txt
@@ -5,3 +5,4 @@ pytest-django==4.5.2
 pytest-cov==4.1.0
 pytest-playwright==0.4.4
 pytest-env==1.1.3
+tzdata==2024.1
